requirements.txt
boto3==1.28.30
botocore==1.31.30
exceptiongroup==1.1.3
iniconfig==2.0.0
jmespath==1.0.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.16
